requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
geoalchemy2==0.14.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
apscheduler==3.10.4
beautifulsoup4==4.12.2